Moved cards to its own module and converted it to a typescript file.
